feat: Implement product search, filtering, and detailed product views
This commit introduces a comprehensive update to the product discovery flow, including full search and filtering capabilities, a detailed product screen, and significant architectural refactoring of use cases and repositories.

Changes include:

- **Product & Home Features:**
  - Implement dynamic product loading with pagination and search in `HomeScreen`.
  - Add `FilterScreen` with options for sorting, price ranges, categories, and stock status.
  - Implement `ProductScreen` with size selection, quantity management, and image carousels.
  - Update `ProductCard` to display average ratings and handle optional price formatting.
  - Integrated `ProductViewModel` and `HomeViewModel` to manage complex UI states and filtering logic.

- **Networking & Data:**
  - Add `ProductApi`, `ReviewApi`, and `CategoryApi` for remote data access.
  - Implement `CategoryRepository` and update `ProductRepository` to support advanced search queries.
  - Refactor `ReviewRepository` and `PaymentMethodRepository` to handle more granular data, such as cash vs. card payment types.
  - Updated DTOs to include product ratings, review counts, and category hierarchies.

- **Domain & Use Cases:**
  - Grouped related use cases (e.g., `ProductUseCases`, `CartUseCases`, `AuthUseCases`, `ReviewUseCases`) into consolidated files for better maintainability.
  - Introduced `CategoryUseCases` for fetching product categories.
  - Updated `AuthRepository` to support refreshing both access and refresh tokens.

- **UI & Navigation:**
  - Update `NavHost` to support parameterized routes for product details (`Product/{productId}`).
  - Enhance `CustomSearchBar` to support external query state management.
  - Refactor `PaymentMethodsSheetContent` to distinguish between credit cards and cash methods.
  - Add `kotlin-parcelize` plugin to support passing complex filter states between screens.

- **General:**
  - Refined string resources for better error messaging and label consistency.
  - Improved DI configuration in `NetworkModule` and `RepositoryModule` for new services.

+++ app/src/main/java/setixx/software/kaimono/domain/model/Review.kt
@@ -3,7 +3,7 @@ package setixx.software.kaimono.domain.model
 data class Review(
     val id: Long,
     val userName: String,
-    val rating: Int,
+    val rating: Short,
     val comment: String?,
     val createdAt: String
 )
@@ -11,11 +11,11 @@ data class Review(
 data class CreateReview(
     val productPublicId: String,
     val orderPublicId: String,
-    val rating: Int,
+    val rating: Short,
     val comment: String?
 )
 
 data class UpdateReview(
-    val rating: Int,
+    val rating: Short,
     val comment: String?
 )
